using the same id for events generated by the same MonitoringEventDispatcher instance
this is done so as to accomodate the use of the library from ObjectService.
Every request/response in ObjectService is handled in its own process so, a new MonitoringEventDispatcher
instance is created each time and a new id is generated for all events of the same instance. To get
events with different id, you should acquire a new MonitoringEventDispatcher instance.